pyproject.toml

[project]
name = "frame"
version = "0.1.0"
description = "A framework for machine learning tasks."
requires-python = ">=3.10,<3.14"
dependencies = [
    "hydra-core>=1.3.2",
    "mlcroissant>=1.0.22",
    "mlflow>=3.6.0",
    "omegaconf>=2.3.0",
    "antlr4-python3-runtime>=4.9.1",
    "pyyaml>=6.0.1",
    "pytorch-lightning>=2.5.6",
    "ruff>=0.14.4",
    "scikit-learn>=1.3.2",
    "segmentation-models-pytorch>=0.3.3",
    "submitit>=1.5.3",
    "torchgeo>=0.6.2",
    "torchx",
    "torchdata==0.9.0",
    "tensorflow-datasets>=4.9.9",
    "apache-beam>=2.69.0",
    "tensorflow-metadata>=1.10",
    "gpytorch>=1.14.2",
]

[tool.ruff]
line-length = 88

[tool.uv.sources]
torchx = { git = "https://github.com/meta-pytorch/torchx.git" }
